+++ packages/spot-consumer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.49](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.48...@injectivelabs/spot-consumer@0.0.49) (2021-07-14)
+
+**Note:** Version bump only for package @injectivelabs/spot-consumer
+
+
+
+
+
 ## [0.0.48](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.47...@injectivelabs/spot-consumer@0.0.48) (2021-06-25)